Fixed issues

+++ core/test/memory.py
@@ -250,10 +250,10 @@ class MemoryTester(Configurable):
         log.info("Launching remotely ...")
 
         # Get the number of free cores
-        ncores = math.floor(self.launcher.single_remote.free_sockets) * self.launcher.single_remote.cores_per_socket
+        ncores = int(math.floor(self.launcher.single_remote.free_sockets)) * self.launcher.single_remote.cores_per_socket
 
         # Determine the number of cores per process, the number of threads per core and thus the number of threads per process
-        cores_per_process = max(math.floor(ncores / self.config.nprocesses), 1)
+        cores_per_process = max(int(math.floor(ncores / self.config.nprocesses)), 1)
         threads_per_core = self.launcher.single_remote.threads_per_core if self.launcher.single_host.use_hyperthreading else 1
         nthreads = cores_per_process * threads_per_core
         ncores = nthreads / threads_per_core * self.config.nprocesses
